gitlab: remove compatibility layer for client <1.2.0 (#7671)
This PR removes the compatibility layer introduced in #7670 to support
both >=v1.2.0 and <=0.159.0.
With the upgrade of master to v1.2.0, we no longer need to keep the
compatibility layer and we can drop typecasts since they are no longer
needed.

Related to #7670
Related to https://github.com/gravitational/teleport/pull/61915


Source-commit: gravitational/teleport.e@1f3a3979e5d373ce69e1e55cb6ea178a4b72f17b
Source-pull-requests: gravitational/teleport.e#7671

+++ e/lib/accessgraph/gitlab/gitlab.go
@@ -97,8 +97,7 @@ func (g *gitlabFetcher) getProjects() (
 		}
 		out = append(out, prj)
 
-		// TODO(tigrato):  remove int64 conversion when gitlab client is updated
-		members, err := g.client.getProjectMembers(int64(project.ID))
+		members, err := g.client.getProjectMembers(project.ID)
 		if err != nil {
 			return nil, nil, trace.Wrap(err)
 		}
@@ -135,8 +134,7 @@ func (g *gitlabFetcher) getGroups() (
 		}
 		out = append(out, grp)
 
-		// TODO(tigrato):  remove int64 conversion when gitlab client is updated
-		members, err := g.client.getGroupMembers(int64(group.ID))
+		members, err := g.client.getGroupMembers(group.ID)
 		if err != nil {
 			return nil, nil, trace.Wrap(err)
 		}
